WITRA files
This repository contains the latest WITRA files as of 24th of December
2014. There are 4 folders. "inc" and "libs" are provided by Toradex and
contain include files for the VS2008 projects. "WITRAv11-FINAL" is the
final version of WITRA's code and "REPEAT_TELEOPERATION" has the code to
read from the datalog and repeat the movements performed by the user on
the last session.